+++ CHANGELOG.md
@@ -16,10 +16,13 @@ This is the official BubbleOS changelog! All features will be recorded in _'grou
 - Changed all references from "folder" to "directory" to help create a consistency of terminology usage.
 - Changed the `sysinfo` CPU information so that the speed will not be outputted if it is zero, and the CPU information will not be shown if BubbleOS is unable to get CPU information.
 - Fixed an issue where the update checker incorrectly considered a full release of the same build as outdated if a release candidate existed. For example, if the installed version was Build 200, but the fetched version was Build 200 Release Candidate 1, the update checker would mistakenly treat the release candidate as newer.
+- Fixed an issue where the `time` command would not show times in AM/PM format by default.
 - Fixed an issue where showing path basenames would show an empty string when at the root of a drive.
 - Fixed an issue where reading an empty directory with `dirtree` would not show a message saying the directory was empty.
 - Fixed an issue where the alphabetical sorting of the `sysinfo` command would sort by words starting with uppercase letters first.
 - Fixed an issue where the `dirtree` command would not show an empty line after showing the tree.
+- Fixed an issue where the `tasklist` and `taskkill` commands would crash if they could not get the process list.
+- Fixed an issue where the `symlink` command crashed if making a symbolic link was unable to be done.
 - Fixed an issue where if the history in the configuration file was not a valid data type, BubbleOS would crash and bootloop instead of fixing the issue automatically.
 
 ### Removed Features